% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terraTCGAdata.R
\name{ExperimentListData}
\alias{ExperimentListData}
\title{Import Terra TCGA data as an ExperimentList}
\usage{
ExperimentListData(
  assayNames,
  sampleCode,
  workspace = terraWorkspace(),
  namespace = .DEFAULT_NAMESPACE,
  tablename = .DEFAULT_TABLENAME,
  verbose = TRUE
)
}
\arguments{
\item{assayNames}{character() A vector of assays selected from the colnames
of \code{getAssayTable}.}

\item{sampleCode}{character(1) The sample code used to filtering samples
e.g., "01" for Primary Solid Tumors, see
\code{data("sampleTypes", package = "TCGAutils")} for reference}

\item{workspace}{character(1) The Terra Data Resources workspace from which
to pull TCGA data (default: see \code{terraWorkspace()}). This is set to a
package-wide option.}

\item{namespace}{character(1) The Terra Workspace Namespace that
defaults to "broad-firecloud-tcga" and rarely needs to be changed.}

\item{tablename}{The Terra data model table from which to extract the
clinical data (default: "sample")}
}
\value{
An \code{ExperimentList} of assays
}
\description{
Import Terra TCGA data as an ExperimentList
}
\examples{

ExperimentListData(
    assayNames = c("protein_exp__mda_rppa_core__mdanderson_org__Level_3__protein_normalization__data",
    "snp__genome_wide_snp_6__broad_mit_edu__Level_3__segmented_scna_minus_germline_cnv_hg18__seg"),
    sampleCode = c("01", "10"),
    workspace = "TCGA_COAD_OpenAccess_V1-0_DATA"
)

}
