% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terraTCGAdata-defunct.R
\name{findTCGAworkspaces}
\alias{findTCGAworkspaces}
\title{findTCGAworkspaces is defunct in terraTCGAdata}
\usage{
findTCGAworkspaces()
}
\value{
Defunct functions return an error
}
\description{
The function has been replaced by `selectTCGAworkspace` which
  provides an interactive workspace selection method.
}
