% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terraTCGAworkspaces.R
\name{terraWorkspace}
\alias{terraWorkspace}
\alias{findTCGAworkspaces}
\title{Obtain or set the Terra Workspace Project Dataset}
\usage{
terraWorkspace(projectName = NULL)

findTCGAworkspaces(project = "^TCGA", cancerCode = ".*")
}
\arguments{
\item{projectName}{character(1) A project code usually in the form of
\verb{TCGA_CODE_OpenAccess_V1-0_DATA}. See \code{findTCGAworkspaces} for a list of
project codes.}
}
\value{
A Terra TCGA Workspace name
}
\description{
Terra allows access to about 71 open access TCGA datasets. A dataset
workspace can be set using the \code{terraWorkspace} function with a \code{projectName}
input. Use the \code{findTCGAworkspaces} function to list all of the available
open access TCGA data workspaces.
}
\section{Functions}{
\itemize{
\item \code{findTCGAworkspaces}: Function to enumerate the available TCGA data
workspaces in Terra
}}

\examples{

findTCGAworkspaces()

}
