% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{ExperimentListData}
\alias{ExperimentListData}
\title{Import Terra TCGA data as an ExperimentList}
\usage{
ExperimentListData(assayNames, sampleCode)
}
\arguments{
\item{assayNames}{character() A vector of assays selected from the colnames
of `getAssayTable`.}

\item{sampleCode}{character(1) The sample code used to filtering samples
e.g., "01" for Primary Solid Tumors, see
`data("sampleTypes", package = "TCGAutils")` for reference}
}
\value{
An `ExperimentList` of assays
}
\description{
Import Terra TCGA data as an ExperimentList
}
\examples{

ExperimentListData(
    assayNames = c("protein_exp__mda_rppa_core__mdanderson_org__Level_3__protein_normalization__data",
    "snp__genome_wide_snp_6__broad_mit_edu__Level_3__segmented_scna_minus_germline_cnv_hg18__seg"),
    sampleCode = c("01", "10")
)

}
