% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terraTCGAdata.R
\name{getAssayData}
\alias{getAssayData}
\title{Obtain assay datasets from Terra}
\usage{
getAssayData(
  assayName,
  sampleCode = "01",
  tablename = "sample",
  metacols = .PARTICIPANT_METADATA_COLS
)
}
\arguments{
\item{assayName}{character() The name of the assay dataset column from
\code{getAssayTable} to import into the current workspace.}

\item{sampleCode}{character(1) The sample code used to filtering samples
e.g., "01" for Primary Solid Tumors, see
\code{data("sampleTypes", package = "TCGAutils")} for reference}

\item{tablename}{The terra data model table from which to extract the
clinical data (default: "sample")}

\item{metacols}{The set of columns that comprise of the metadata columns.
See the \code{.PARTICIPANT_METADATA_COLS} global variable}
}
\value{
Either a matrix or RaggedExperiment depending on the assay selected
}
\description{
Obtain assay datasets from Terra
}
\examples{

getAssayData(
    assayName = "protein_exp__mda_rppa_core__mdanderson_org__Level_3__protein_normalization__data",
    sampleCode = c("01", "10"),
    tablename = "sample"
)

getAssayData(
    assayName = "snp__genome_wide_snp_6__broad_mit_edu__Level_3__segmented_scna_minus_germline_cnv_hg18__seg",
    sampleCode = c("01", "10"),
)

getAssayData(
    assayName = "snp__genome_wide_snp_6__broad_mit_edu__Level_3__segmented_scna_minus_germline_cnv_hg18__seg"
)

}
\seealso{
\code{\link[=getAssayTable]{getAssayTable()}}
}
