% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terraTCGAdata.R
\name{getAssayTable}
\alias{getAssayTable}
\title{Obtain a reference table for assay data in the Terra data model}
\usage{
getAssayTable(tablename = "sample", metacols = .PARTICIPANT_METADATA_COLS)
}
\arguments{
\item{tablename}{The terra data model table from which to extract the
clinical data (default: "sample")}

\item{metacols}{The set of columns that comprise of the metadata columns.
See the \code{.PARTICIPANT_METADATA_COLS} global variable}
}
\value{
A tibble of pointers to resources within the Terra data model
}
\description{
The column names in the output can be used in the `getAssayData` function.
}
