% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{getClinical}
\alias{getClinical}
\title{Obtain clinical data}
\usage{
getClinical(
  tablename = "sample",
  participants = TRUE,
  columnName = NULL,
  metacols = .PARTICIPANT_METADATA_COLS
)
}
\arguments{
\item{tablename}{The terra data model table from which to extract the
clinical data (default: "sample")}

\item{participants}{logical(1) Whether to merge the participant table
from `avtable("participant")` to the clinical data}

\item{columnName}{The name of the column to extract files, see
`getClinicalTable` table.}

\item{metacols}{The set of columns that comprise of the metadata columns.
See the `.PARTICIPANT_METADATA_COLS` global variable}
}
\value{
A tibble of Google Storage resource locations e.g.,
    `gs://firecloud...`
}
\description{
Obtain clinical data
}
\examples{

getClinical()

}
