% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{getClinical}
\alias{getClinical}
\title{Get the list of clinical datasets}
\usage{
getClinical(tablename = "sample", metacols = .PARTICIPANT_METADATA_COLS)
}
\arguments{
\item{tablename}{The terra data model table from which to extract the
clinical data (default: "sample")}

\item{metacols}{The set of columns that comprise of the metadata columns.
See the `.PARTICIPANT_METADATA_COLS` global variable}
}
\value{
A tibble of Google Storage resource locations e.g.,
    `gs://firecloud...`
}
\description{
Get the list of clinical datasets
}
\examples{

getClinical()

}
