% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical.R
\name{getClinical}
\alias{getClinical}
\title{Obtain clinical data}
\usage{
getClinical(
  columnName,
  participants = TRUE,
  tablename = "sample",
  metacols = .PARTICIPANT_METADATA_COLS
)
}
\arguments{
\item{columnName}{The name of the column to extract files, see
\code{getClinicalTable} table.}

\item{participants}{logical(1) Whether to merge the participant table
from \code{avtable("participant")} to the clinical data}

\item{tablename}{The terra data model table from which to extract the
clinical data (default: "sample")}

\item{metacols}{The set of columns that comprise of the metadata columns.
See the \code{.PARTICIPANT_METADATA_COLS} global variable}
}
\value{
A tibble of Google Storage resource locations e.g.,
\verb{gs://firecloud...}
}
\description{
Obtain clinical data
}
\examples{

getClinical()

}
