% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical.R
\name{getClinicalTable}
\alias{getClinicalTable}
\title{Obtain the reference table for clinical data}
\usage{
getClinicalTable(tablename = "sample", metacols = .PARTICIPANT_METADATA_COLS)
}
\arguments{
\item{tablename}{The terra data model table from which to extract the
clinical data (default: "sample")}

\item{metacols}{The set of columns that comprise of the metadata columns.
See the \code{.PARTICIPANT_METADATA_COLS} global variable}
}
\value{
A tibble of Google Storage resource locations e.g.,
\verb{gs://firecloud...}
}
\description{
The column names in the output table can be used in the \code{getClinical}
function.
}
