% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical.R
\name{sampleTypesTable}
\alias{sampleTypesTable}
\title{Get an overview of the samples available in the workspace}
\usage{
sampleTypesTable()
}
\value{
A \code{tibble} of sample codes and frequency along with their
definition and short letter code
}
\description{
The function provides an overview of samples from the \code{avtables("sample")}
table for the current workspace. Along with the sample codes and frequencies,
the output provides a description for each code and the short letter codes.
}
\examples{

sampleTypesTable()

}
