% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{terraTCGAdata}
\alias{terraTCGAdata}
\title{Obtain a MultiAssayExperiment from the Terra workspace}
\usage{
terraTCGAdata(assays, sampleCode = NULL, split = TRUE)
}
\arguments{
\item{assays}{character() A character vector of assay names taken from
\code{getAssayTable()}}

\item{sampleCode}{character() A character vector of sample codes from
\code{sampleTypesTable()}}

\item{split}{logical(1L) Whether or not to split the \code{MultiAssayExperiment}
by sample types using \code{splitAssays} helper function.}
}
\value{
A \code{MultiAssayExperiment} object with n number of assays corresponding
to the \code{assays} argument.
}
\description{
Workspaces on Terra come pre-loaded with TCGA Data. The examples in the
documentation correspond to the TCGA_COAD_OpenAccess_V1 workspace that
can be found on \url{app.terra.bio}.
}
\examples{

terraTCGAdata(
    assays = c("protein_exp__mda_rppa_core__mdanderson_org__Level_3__protein_normalization__data",
    "rnaseqv2__illuminahiseq_rnaseqv2__unc_edu__Level_3__RSEM_genes_normalized__data"),
    sampleCode = NULL,
    split = FALSE
)

}
