% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terraTCGAworkspaces.R
\name{terraTCGAworkspace}
\alias{terraTCGAworkspace}
\alias{findTCGAworkspaces}
\title{Obtain or set the Terra Workspace Project Dataset}
\usage{
terraTCGAworkspace(projectName = NULL)

findTCGAworkspaces(project = "^TCGA", cancerCode = ".*")
}
\arguments{
\item{projectName}{character(1) A project code usually in the form of
\verb{TCGA_CODE_OpenAccess_V1-0_DATA}. See \code{findTCGAworkspaces} for a list of
project codes.}
}
\value{
A Terra TCGA Workspace name
}
\description{
Terra allows access to about 71 open access TCGA datasets. A dataset
workspace can be set using the \code{terraTCGAworkspace} function with a \code{projectName}
input. Use the \code{findTCGAworkspaces} function to list all of the available
open access TCGA data workspaces.
}
\details{
\preformatted{Note that GDC workspaces are not supported and are excluded
from the search results. GDC workspaces use a Terra workflow to download
TCGA data rather than providing Google Bucket storage locations for easy
data retrieval.
}
}
\section{Functions}{
\itemize{
\item \code{findTCGAworkspaces}: Function to enumerate the available TCGA data
workspaces in Terra
}}

\examples{

findTCGAworkspaces()

}
