% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terraTCGAworkspaces.R
\name{terraTCGAworkspace}
\alias{terraTCGAworkspace}
\alias{findTCGAworkspaces}
\title{Obtain or set the Terra Workspace Project Dataset}
\usage{
terraTCGAworkspace(projectName = NULL)

findTCGAworkspaces(project = "^TCGA", cancerCode = ".*")
}
\arguments{
\item{projectName}{character(1) A project code usually in the form of
\verb{TCGA_CODE_OpenAccess_V1-0_DATA}. See \code{findTCGAworkspaces} for a list of
project codes.}

\item{project}{character(1) A prefix for the regex search across all public
projects on the terra platform (default: \code{"^TCGA"}). Usually, this does not
change.}

\item{cancerCode}{character(1) Corresponds to the TCGA cancer code (e.g,
"ACC" for AdrenoCortical Carcinoma) of interest. The default value of
(\verb{.*}) provides all available cancer datasets.}
}
\value{
A Terra TCGA Workspace name
}
\description{
Terra allows access to about 71 open access TCGA datasets. A dataset
workspace can be set using the \code{terraTCGAworkspace} function with a
\code{projectName} input. Use the \code{findTCGAworkspaces} function to list all
of the available open access TCGA data workspaces.
}
\details{
Note that GDC workspaces are not supported and are excluded
from the search results. GDC workspaces use a Terra workflow to download
TCGA data rather than providing Google Bucket storage locations for easy
data retrieval. To reset the \code{terraTCGAworkspace}, use
\code{terraTCGAworkspace(NULL)} and you will be prompted to select from a list
of TCGA workspaces. You may also check the current active workspace by
running \code{terraTCGAworkspace()} without any inputs.
}
\section{Functions}{
\itemize{
\item \code{findTCGAworkspaces()}: Function to enumerate the available TCGA data
workspaces in Terra

}}
\examples{
if (AnVIL::gcloud_exists())
  findTCGAworkspaces()

}
